config/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0
